[dev.typeparams] cmd/compile/internal/types2: handle untyped constant arithmetic overflow
Factor out the existing "costant representation" check after
untyped constant arithmetic and combine with an overflow check.

Use a better heuristic for determining the error position if we
know the error is for a constant operand that is the result of an
arithmetic expression.

Minor related cleanups.

Without this change, untyped constant arithmetic could lead to
arbitrarily large (integer) constants, limited in size only by
space and time.

Change-Id: Id3cea66c8ba697ff4c7fd1e848f350d9713e3c75