[dev.regabi] cmd/link: use unwrapped runtime.addmoduledata as the primary symbol
runtime.addmoduledata is referenced in init code called by the
dynamic linker. It should not be called via a wrapper. When
linking against a runtime in a shared library, the definition
has mangled name but the reference has unmangled name. This CL
changes it to not mangle the "real" runtime.addmoduledata symbol,
so the reference works.

I'm pretty sure this is not the right thing to do. But it does
make shared library tests working when the g register is used
(next CL). We need better story for symbol ABI resolution for
shared linkage.

Change-Id: I2165f149bc83d513e81eb1eb4ee95464335b0e75