archive/tar: strip leading 0s in decimal length for a PAX record
The specification under section "pax Extended Header" at

    https://pubs.opengroup.org/onlinepubs/9699919799/utilities/pax.html#tag_20_92_13_03

requires that a PAX record be of the format:

    "%d %s=%s\n", <length>, <keyword>, <value>

and specifies that:

    The <length> field shall be the decimal length of the extended
    header record in octets, including the trailing <newline>.

However, by the definition of a decimal length, it could be
0-padded thus, the remedy is to strip the length's leading 0's, and with
that remove the leading "00" so that if for example "0" were passed in,
it would be recognized as a plain 0.
Added test cases to ensure that the parsing still proceeds as expected.

The prior crashing repro:

    0000000000000000000000000000000030 mtime=1432668921.098285006\n30 ctime=2147483649.15163319

exposed a fallacy in the code, that assumed that the length would ALWAYS be a
non-padded decimal length string.

This bug has existed since Go1.1 as per CL 6700047.
Thanks to Josh Bleecher Snyder for fuzzing this package.

Fixes #40196

Change-Id: I32e0af4887bc9221481bd9e8a5120a79f177f08c