[dev.regabi] cmd: add ABICpp as possible ABI for assembly functions
DO NOT SUBMIT

[Experimental CL, intended for demonstration purposes. Needs evaluation.
 also note that this does not currently pass all.bash.]

Change the assembler to allow assembly functions to be defined as
"ABICpp". This ABI tells the compiler that the function in question
uses the platform C/C++ ABI, and is intended to be called from C code
as opposed to Go code; such functions don't need ABI wrappers.

Previously the way we avoided the introduction of a wrapper was to
define such functions as ABIInternal, however given that they don't
actually use the internal Go ABI, this was somewhat misleading.

Change-Id: I4304f613c1477fd3ca253db4e28bbac53de84646