[dev.regabi] runtime: take address in assembly for syscall wrappers on darwin
CL 288092 marked the syscall wrappers as ABIInternal, because we
need to call to them without an ABI wrapper, and they are
referenced in Go using funcPC.

This CL is an alternative approach: As the Go side only needs
their addresses, we can take address directly in assembly, and
only pass the address to Go. Then we don't need any ABI marking.
In fact, the wrappers can be file-local, as they are no longer
referenced in Go side.

This is a proof-of-concept change. It does only the runtime
package now. If this approach is preferred, I'll go ahead
changing the syscall package as well.

DO NOT SUBMIT.

For #44065.

Change-Id: Id25091690f0edef186ac4c3c8bede2b7d8a82d6c