[dev.boringcrypto.go1.16] all: branch dev.boringcrypto.go1.16
We don't set a parent-branch because we can't use "git-codereview
sync-branch" to do our merges since we want them to happen at a specific
commit (the release tag), not at the HEAD of the release-branch.

We have misc/boring/merge.sh for release merges.

To be fair, we can't use it for dev.boringcrypto merges either, because
those also need to happen at the first conflicting commit, not at HEAD,
but I am hoping eventually we'll get release merges into the release
process, and maybe add a flag to sync-branch for master merges.

Change-Id: Ia1fbe284f0d7fe82c29109b40601139697e2e1e3