cmd/gofmt: atomically swap the new file
This change modifies `gofmt -w` to replace the file atomically on POSIX
filesystems. Instead of creating a backup of the original file and then
restoring it if it fails, do all the procedure in a temporary file and then
swap it out with the original.

Change-Id: Id37fb3099a250f0596316eda08a17f740a05b51b
GitHub-Last-Rev: 6e4f7bddc64d8fdd64bf61b777b73557f9d16a5c
GitHub-Pull-Request: golang/go#44173