[dev.fuzz] testing: move inFuzzFn checks from common to F
inFuzzFn is set when the fuzz function is called. While it's set,
F methods that have side effects like Skip and Fail may not be called.

Previously, inFuzzFn was in common, and we checked it in the common
implementation of those methods. This causes problems in CL 290693 for
recursive methods like common.Fail. If T.Fail is called by the fuzz
function, it calls common.Fail on the parent F's common. That should
not panic.

Change-Id: I841b12f77d9c77f5021370d03313e71b4ef50102