[dev.typeparams] go/types: conversions to type parameters are not constant
This is a port of CL 290471 to go/types. Notably, this change preserves
the existing check for constant types in recordTypeAndValue, which uses
is(..., isConstType) rather than the isConstType predicate. I believe
types2 should do this as well: currently types2 is avoiding a panic only
because it erroneously records the type as the first member of its type
list.

Longer term, we should consider allowing const type parameters if the
optype is a sum of constant types.

Change-Id: Iaa91ffa740b5f08a5696bd96918a866bffd7aef6