runtime/cgo: implement for windows/arm and windows/arm64
DO NOT SUBMIT

This is a work in progress. There's still work to be done inside the
linker apparently. Feel free to take this skeleton and as the basis for
a new commit and discard this one, if it's helpful. Status: it appears
to be working on windows/arm.

This CL is part of a stack adding windows/arm64
support (#36439), intended to land in the Go 1.17 cycle.

Fixes #43800.

Change-Id: If9651cf04576dfd183e4706dd4dfe343cee3414e