os/exec: don't wait if grandchild holds pipe open
If a pipe is used to read standard output or standard error from a
child process, and the child process passes the pipe to a grandchild
process, we used to wait until the grandchild process closed the pipe
(typically this meant waiting for the grandchild process to exit).

With this change, we stop waiting once the child process exits,
even if a grandchild process is keeping the pipe open.

Fixes #23019

Change-Id: I3eb6d46e7e8e553d87e1e4357a8e5a7d9a295ea4