[dev.typeparams] go/types: use a new ast.ListExpr for multi-type instances
Modify go/parser to consistently represent type instantiation as an
ast.IndexExpr, rather than use an ast.CallExpr (with Brackets:true) for
instantiations with multiple type parameters. To enable this, introduce
a new ast expr type: ListExpr.

This brings go/types in line with types2. In a subsequent CL, call.go
and expr.go will be marked as reviewed.

Change-Id: I30cd0181c7608f1be7486a9a8b63df993b412e85