cmd/asm: move riscv64 jump instruction encoding tests together
Adding or removing any instructions between the start label and the
jump instruction tests causes the jump instruction tests to start
failing since their offset, and therefore encoding, changes. To
reduce the likelihood of future changes encountering this issue I've
moved all of the jump instruction tests to the start of the function.

This should allow us to add and remove instructions from the rest
of the test function without needing to fixup the jump tests.

Change-Id: I1b1e8d824834e0c1979927ffb20629108eff591d